refactor: introduce testutil.h (deduplicate `random_fe_`, `ge_equals_` helpers)
